Update to 1.4.50, add confirmation to update script

+++ update.sh
@@ -10,22 +10,31 @@ set -e    # Exit on first failure
 PKG_NAME='sublime-botocore'
 
 DEP_REPO='boto/botocore'
-DEP_RELEASE='1.4.49'
+DEP_RELEASE="${1:-1.4.50}"
 # ----- #
+
+echo "${PURPLE}Update ${DEP_REPO} to ${DEP_RELEASE}${RESET}"
+
+# Confirmation
+echo ''
+echo 'Continuing will download a new release and replace the entire "all" directory.'
+
+read -p "${RED}Are you sure? [y/N]:${RESET} " confirm
+[[ ! "$confirm" =~ ^[Yy]$ ]] && exit 1
 
 tarfile="${PKG_NAME}-${DEP_RELEASE}.tar.gz"
 
-# Download release tarball
+# Download a release tarball
 curl -v --output "/tmp/${tarfile}" --location -- "https://github.com/${DEP_REPO}/archive/${DEP_RELEASE}.tar.gz"
 
-## ALT: Clone detached head from a Github tagged release
+## ALT: Clone a detached head from a Github tagged release
 #git clone --branch="$DEP_RELEASE" --single-branch --depth 1 "${DEP_REPO_URL}" all
 #rm -rf 'all/.git'
 
-# Delete existing dependency files
+# Delete existing files
 [ -d 'all' ] && rm -r 'all/'
 
-# Create destination dir and extract the contents of the archive
+# Create "all" dir and extract tarball contents
 mkdir 'all'
 tar -zxvf "/tmp/${tarfile}" -C 'all' --strip-components 1
